+++ src/main/java/com/example/model/Question.java
@@ -1,7 +1,7 @@
 package com.example.model;
 
 public class Question {
-    private Integer id;
+    private Long id;
 
     private String title;
 
@@ -11,7 +11,7 @@ public class Question {
 
     private Long gmtModified;
 
-    private Integer creator;
+    private Long creator;
 
     private Integer commentCount;
 
@@ -21,11 +21,11 @@ public class Question {
 
     private String tag;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -61,11 +61,11 @@ public class Question {
         this.gmtModified = gmtModified;
     }
 
-    public Integer getCreator() {
+    public Long getCreator() {
         return creator;
     }
 
-    public void setCreator(Integer creator) {
+    public void setCreator(Long creator) {
         this.creator = creator;
     }
